+++ defineTest.js
@@ -7,14 +7,21 @@ export default function defineTest(
   transformName,
   options,
   testFilePrefix,
-  testOptions
+  testOptions = {}
 ) {
   testFilePrefix = Array.isArray(testFilePrefix)
     ? path.join(...testFilePrefix)
     : testFilePrefix;
   const fixtureDir = path.join(dirName, "..", "__testfixtures__");
-  const inputPath = path.join(fixtureDir, testFilePrefix + `.input.js`);
-  const outputPath = path.join(fixtureDir, testFilePrefix + `.output.js`);
+  const fileExtension = testOptions.parser || "js";
+  const inputPath = path.join(
+    fixtureDir,
+    testFilePrefix + `.input.` + fileExtension
+  );
+  const outputPath = path.join(
+    fixtureDir,
+    testFilePrefix + `.output.` + fileExtension
+  );
   const source = fs.readFileSync(inputPath, "utf8");
   const expectedOutput = fs.readFileSync(outputPath, "utf8");
   const transform = require(path.join(dirName, "..", transformName));