+++ transforms/add-prop/__testfixtures__/typescript/BooleanTrue.input.tsx
@@ -1,0 +1,7 @@
+import Button from "carbon-react/lib/components/button";
+
+interface SomeInterface {
+  label: string;
+}
+
+export const asTrue = (arg: string): JSX.Element => <Button />