+++ transforms/replace-row-column-with-grid/__testfixtures__/WithPropsSpreadFromDeclarationScope.input.js
@@ -1,0 +1,19 @@
+import { Row, Column } from "carbon-react/lib/components/row";
+
+const WithPropsSpreadFromDeclarationScope = () => {
+  const rowProps = {
+    gutter: "medium",
+    columns: 5,
+    columnDivide: true,
+  };
+
+  return (
+    <Row {...rowProps}>
+      <Column columns={rowProps.columns}>col1</Column>
+      <Column columns={rowProps.columns}>col2</Column>
+      <Column columns={rowProps.columns}>col3</Column>
+    </Row>
+  );
+}
+
+export default WithPropsSpreadFromDeclarationScope;